Move issue milestone assign to issue service and move webhook to notification
